refactor: extract conversation viewer document template
